Added some stats and moved formatting to it's own file
